Job Recommender MCP/requirements.txt
streamlit
openai
apify-client
python-dotenv
supabase
weasyprint
fitz
PyMuPDF